+++ CHANGELOG.md
@@ -24,7 +24,7 @@ each for its own `findOne*` or `find*` methods
 * `SpecificRepository` is deprecated for now
 * `transaction` method has been removed from `Repository`. Use `EntityManager#transaction` method instead
 * custom repositories do not support container anymore
-* added ActiveRecord support (by extending EntityModel) class
+* added ActiveRecord support (by extending BaseEntity) class
 * controller / subscriber / migrations from options tsconfig now appended with a project root directory
 * removed naming strategy decorator, naming strategy by name functionality. 
 Now naming strategy should be registered by passing naming strategy instance directly